move deck slots for LLD test labware
