+++ hardware-testing/hardware_testing/liquid_sense/__main__.py
@@ -35,9 +35,9 @@ from .report import build_ls_report, store_config, store_serial_numbers
 from .post_process import process_csv_directory, process_google_sheet
 
 from hardware_testing.protocols.liquid_sense_lpc import (
-    liquid_sense_ot3_p50_single,
+    liquid_sense_ot3_p50_single_vial,
     liquid_sense_ot3_p50_multi,
-    liquid_sense_ot3_p1000_single,
+    liquid_sense_ot3_p1000_single_vial,
     liquid_sense_ot3_p1000_multi,
     liquid_sense_ot3_p1000_96,
 )
@@ -61,11 +61,11 @@ MAX_PROBE_SECONDS = 3.5
 
 LIQUID_SENSE_CFG = {
     50: {
-        1: liquid_sense_ot3_p50_single,
+        1: liquid_sense_ot3_p50_single_vial,
         8: liquid_sense_ot3_p50_multi,
     },
     1000: {
-        1: liquid_sense_ot3_p1000_single,
+        1: liquid_sense_ot3_p1000_single_vial,
         8: liquid_sense_ot3_p1000_multi,
         96: liquid_sense_ot3_p1000_96,
     },
